ray: Mount host's shared memory
Before, the containers used the default shared memory size of 64M,
resulting in the following error when running the Ray examples:

There is not enough space to create this object, so evicting 0 objects to free up 0 bytes.
pyarrow.lib.PlasmaStoreFull: object does not fit in the plasma store

This commit makes it so the containers mount the host machine's shared
memory. On the `rayExample.js` machines (m4.larges), this is 3.9G, which
is enough for the examples to run successfully.